#150: Create localised conftest files
- Local to each backend, with a shared file in test/
